Replace ints with longs (#28)

+++ app/src/commonMain/kotlin/it/vercruysse/lemmyapi/v0x18/datatypes/CommunityAggregates.kt
@@ -4,15 +4,15 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 internal data class CommunityAggregates(
-    val id: Int,
+    val id: Long,
     val community_id: CommunityId,
-    val subscribers: Int,
-    val posts: Int,
-    val comments: Int,
+    val subscribers: Long,
+    val posts: Long,
+    val comments: Long,
     val published: String,
-    val users_active_day: Int,
-    val users_active_week: Int,
-    val users_active_month: Int,
-    val users_active_half_year: Int,
-    val hot_rank: Int,
+    val users_active_day: Long,
+    val users_active_week: Long,
+    val users_active_month: Long,
+    val users_active_half_year: Long,
+    val hot_rank: Long,
 )
