+++ app/src/commonMain/kotlin/it/vercruysse/lemmyapi/v0x18/datatypes/CommunityBlockId.kt
@@ -1,3 +1,3 @@
 package it.vercruysse.lemmyapi.v0x18.datatypes
 
-typealias CommunityBlockId = Int
+typealias CommunityBlockId = Long
