+++ app/src/commonMain/kotlin/it/vercruysse/lemmyapi/v0x18/datatypes/GetReportCountResponse.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 internal data class GetReportCountResponse(
     val community_id: CommunityId? = null,
-    val comment_reports: Int,
-    val post_reports: Int,
-    val private_message_reports: Int? = null,
+    val comment_reports: Long,
+    val post_reports: Long,
+    val private_message_reports: Long? = null,
 )
