+++ app/src/commonMain/kotlin/it/vercruysse/lemmyapi/v0x18/datatypes/PostView.kt
@@ -15,5 +15,5 @@ internal data class PostView(
     val read: Boolean,
     val creator_blocked: Boolean,
     val my_vote: Int = 0,
-    val unread_comments: Int,
+    val unread_comments: Long,
 )
