+++ app/src/commonMain/kotlin/it/vercruysse/lemmyapi/v0x18/datatypes/RemoveCommunity.kt
@@ -7,6 +7,6 @@ internal data class RemoveCommunity(
     val community_id: CommunityId,
     val removed: Boolean,
     val reason: String? = null,
-    val expires: Int? = null,
+    val expires: Long? = null,
     val auth: String,
 )
